A bunch of fixes to make this work with Apache Superset and Pinot
Summary: Added support for:
- Multiple aggregates
- Pinot does not support quoted column names very well. For example
'select "foo", count(*) as "count" group by "foo"' was NPE'ing for me
- Pinot aggregate (with or without group-by) does not support
order-by-limit. It only supports TOP keyword. Further, most group-by's
in pinot default to top 10. Force that to change to top 100.
- support for weird (non SQL compliant) behavior of pinot's multiple
metrics with group-by-top, wherein the limit is applied !independantly!
for each metric.

This was tested with the SQL being spit out from Apache Superset.

There is some hackery in the replacing of orderby-limit with Top for
aggregate queries. In particular: Its hard to figure out whether the
query is an aggregate query without knowing the semantics of the
expression used. So we rely on a hint from the query preparer. I hacked
that hint using an ugly getattr.

In addition some logging and more type support